# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Glory Kamples API",
    "description": "API Kamples — Rust + Axum + OpenAPI",
    "contact": {
      "name": "1ndoryu"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/busqueda/rapida": {
      "get": {
        "tags": [
          "search"
        ],
        "operationId": "legacy_quick_search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Texto a buscar. Mantiene el contrato legacy del dropdown rápido",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respuesta compatible con el endpoint legacy /busqueda/rapida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyQuickSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/search": {
      "get": {
        "tags": [
          "search"
        ],
        "operationId": "global_search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Texto a buscar. Con menos de 2 caracteres devuelve listas vacías",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "types",
            "in": "query",
            "description": "CSV opcional para filtrar tipos: samples,users,collections,songs",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Búsqueda global agrupada por tipo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Estructura de respuesta de error expuesta en la API",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Tipo de error (`not_found`, `unauthorized`, etc.)"
          },
          "message": {
            "type": "string",
            "description": "Mensaje legible para el usuario"
          }
        }
      },
...
      "GlobalSearchResponse": {
        "type": "object",
        "required": [
          "q",
          "samples",
          "users",
          "collections",
          "songs"
        ],
        "properties": {
          "collections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchCollectionResult"
            }
          },
          "q": {
            "type": "string"
          },
          "samples": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchSampleResult"
            }
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchSongResult"
            }
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchUserResult"
            }
          }
        }
      },
...
      "LegacyQuickSearchCollectionResult": {
        "type": "object",
        "required": [
          "id",
          "nombre",
          "slug",
          "totalSamples",
          "creador"
        ],
        "properties": {
          "creador": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "portadaUrl": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string"
          },
          "totalSamples": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "LegacyQuickSearchRelationResult": {
        "type": "object",
        "required": [
          "id",
          "fuente",
          "destino"
        ],
        "properties": {
          "destino": {
            "$ref": "#/components/schemas/LegacyQuickSearchRelationSide"
          },
          "fuente": {
            "$ref": "#/components/schemas/LegacyQuickSearchRelationSide"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "LegacyQuickSearchRelationSide": {
        "type": "object",
        "required": [
          "titulo",
          "slug",
          "artista"
        ],
        "properties": {
          "artista": {
            "type": "string"
          },
          "imagenUrl": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          }
        }
      },
      "LegacyQuickSearchResponse": {
        "type": "object",
        "required": [
          "canciones",
          "samples",
          "sampleos",
          "usuarios",
          "colecciones",
          "todos"
        ],
        "properties": {
          "canciones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LegacyQuickSearchSongResult"
            }
          },
          "colecciones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LegacyQuickSearchCollectionResult"
            }
          },
          "sampleos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LegacyQuickSearchRelationResult"
            }
          },
          "samples": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LegacyQuickSearchSampleResult"
            }
          },
          "todos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LegacyQuickSearchTodoItem"
            }
          },
          "usuarios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LegacyQuickSearchUserResult"
            }
          }
        }
      },
      "LegacyQuickSearchSampleCreator": {
        "type": "object",
        "required": [
          "username",
          "nombreVisible"
        ],
        "properties": {
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "nombreVisible": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "LegacyQuickSearchSampleResult": {
        "type": "object",
        "required": [
          "id",
          "titulo",
          "slug",
          "creador"
        ],
        "properties": {
          "creador": {
            "$ref": "#/components/schemas/LegacyQuickSearchSampleCreator"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "imagenUrl": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          }
        }
      },
      "LegacyQuickSearchSongResult": {
        "type": "object",
        "required": [
          "id",
          "titulo",
          "slug",
          "artistaNombre",
          "totalSampleada"
        ],
        "properties": {
          "artistaNombre": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "imagenUrl": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          },
          "totalSampleada": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "LegacyQuickSearchTodoItem": {
        "type": "object",
        "required": [
          "tipo",
          "score",
          "datos"
        ],
        "properties": {
          "datos": {
            "type": "object"
          },
          "score": {
            "type": "number",
            "format": "float"
          },
          "tipo": {
            "type": "string"
          }
        }
      },
      "LegacyQuickSearchUserResult": {
        "type": "object",
        "required": [
          "id",
          "username",
          "nombreVisible",
          "verificado",
          "totalSeguidores"
        ],
        "properties": {
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombreVisible": {
            "type": "string"
          },
          "totalSeguidores": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "verificado": {
            "type": "boolean"
          }
        }
      },
...
      "SampleCreatorSummary": {
        "type": "object",
        "description": "Resumen público del creador incluido en `GET /api/samples`.",
        "required": [
          "id",
          "username",
          "verificado"
        ],
        "properties": {
          "avatar_url": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre_visible": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string"
          },
          "verificado": {
            "type": "boolean"
          }
        }
      },
...
      "SearchCollectionOwnerSummary": {
        "type": "object",
        "required": [
          "username",
          "nombre_visible"
        ],
        "properties": {
          "nombre_visible": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "SearchCollectionResult": {
        "type": "object",
        "required": [
          "id",
          "nombre",
          "total_samples",
          "creador"
        ],
        "properties": {
          "creador": {
            "$ref": "#/components/schemas/SearchCollectionOwnerSummary"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "portada_url": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "total_samples": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SearchSampleResult": {
        "type": "object",
        "required": [
          "id",
          "titulo",
          "slug",
          "creador"
        ],
        "properties": {
          "creador": {
            "$ref": "#/components/schemas/SampleCreatorSummary"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "imagen_url": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          }
        }
      },
      "SearchSongResult": {
        "type": "object",
        "required": [
          "id",
          "titulo",
          "slug",
          "artista_nombre",
          "total_sampleada"
        ],
        "properties": {
          "artista_nombre": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "imagen_url": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          },
          "total_sampleada": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "SearchUserResult": {
        "type": "object",
        "required": [
          "id",
          "username",
          "nombre_visible",
          "verificado",
          "total_seguidores"
        ],
        "properties": {
          "avatar_url": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre_visible": {
            "type": "string"
          },
          "total_seguidores": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "verificado": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}